Update MousePanFix.lua
This pretty much will now disable from dragging the mouse to pan the camera in game. It is at least somewhat functional now. It'll prevent the game from getting stuck in pan mode by preventing panning with the mouse altogether. I'm going to keep trying to get this to work better but I can guarantee this should keep it from getting stuck trying to pan the camera when you come back from checking that ffxi guide on bg-wiki or ffxiah.com. I'm just not sure how useful it'll be to anyone that utilizes panning the camera.