Add files via upload
I altered the ui element of pointwatch. now it'll only display the relevant gain/hr for you or it should. I included a backup in the BU folder of the original in case there's any issue. also, I added the commands //pw setfont and //pw setsize to allow you to change the font and the font size dynamically.